[ADD] l10n_be_reports_post_wizard: add XML when posting tax return entry
This new module replaces the POST button on tax return entries and opens
the view_account_financial_report_export wizard first, in order to
configure the XML VAT return export file, which is generated along with
the PDF file, and attached in the chatter of the entry.

This module is intended to be merged into l10n_be_reports in a future
version.

OPW-2392054

closes odoo/enterprise#24676

Signed-off-by: Quentin De Paoli <qdp@odoo.com>