[IMP] l10n_be_hr_payroll: add wizard to set employee language
Previously if the language set on the address of an employee wasn't
correct (for belgian documents) an error would be thrown at the user
telling them to change the language.
After this commit, a wizard will be opened instead, allowing to change
the language of the problematic employees directly.

Task ID: 2582516

closes odoo/enterprise#19364

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>