[FIX] account_reports: remove need for t-raw in formula debug popup
The formula was enriched with bespoke HTML on the server-side in order
to link to its dependencies (if possible). This, then, required
injecting the HTML'd formula as-is into the output.

Update `get_formula_popup` such that the formula is instead returned
as a sequence of tokens, describing either literal text or what to
link the item to, and convert to HTML proper during straightforward
template rendering.