[IMP] web_enterprise: drawer-like mobile hamburger menu
Smaller screens use a custom BurgerMenu which simply opens fullscreen.

As mobile users are more accustomed to drawer-like menu, this commit
modernize it a little to match more closely the UX in mobile apps (cf.
current Gmail app on Android).

The following improvements are implemented in this commit:

Hamburger menu doesn't cover the whole screen and a backdrop darken
the current screen to provide a sense of elevation, disconnecting less the
user from its current workflow.

Hamburger sub-menus are unfolded instead of being hidden behind a dropdown,
providing a more direct access (one less tap/click) to those menu items.

Finally, to match the standard behavior on mobile apps, the hamburger
menu can be closed by both clicking/tapping on the backdrop or by
swiping it toward the anchoring screen edge.

Note: while the actual MenuDropdown component is removed to simplify the
implementation (and because their behavior is no longer needed), but the
CSS styling classes are kept to be able to follow the styling applied in
desktop and inherit from it instead of reimplementing it specifically.

Reference:
https://material.io/components/navigation-drawer

Task ID: 2439062

closes odoo/enterprise#19371

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>