[FIX] delivery_ups: convert package dimensions
When package's UoM is different from dimensions' UoM used in UPS
request, the values are incorrect

To reproduce the error:
1. Sales > Configuration > Shipping Methods
2. Open UPS US > UPS Package/customer supplied
    - Dimensions are 10 x 10 x 10 meters
3. Create a SO
4. Add Shipping: UPS US

Error: Here is part of the request sent to the UPS server:
```xml
<ns0:Package>
    <ns0:PackagingType>
        <ns0:Code>02</ns0:Code>
    </ns0:PackagingType>
    <ns0:Dimensions>
        <ns0:UnitOfMeasurement>
            <ns0:Code>IN</ns0:Code>
        </ns0:UnitOfMeasurement>
        <ns0:Length>10</ns0:Length>
        <ns0:Width>10</ns0:Width>
        <ns0:Height>10</ns0:Height>
    </ns0:Dimensions>
```
The dimensions are 10 x 10 x 10 inches, which are incorrect. The values
must be converted.

This commit allows the conversion to `CM` and `IN` since these are the
two UoM available:
https://github.com/odoo/enterprise/blob/cb13edb79209f15409734488d499e78158a3ba08/delivery_ups/models/delivery_ups.py#L44

OPW-2523952

closes odoo/enterprise#19412

X-original-commit: 4801d06ba7d5901e0ee95aa5f7abf3e13dde506f
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>