[IMP] timesheet_grid: add progress bar in task gantt view
Purpose of the task is to display the progress bar in gantt view
so user see the progress of task from gantt view.

So in this commit, display the progressbar in pill of the task
and also added the total_hours_spent (or effective_hours if sub-tasks
are not enabled on the project) field to the task popover.

closes odoo/enterprise#19569

Taskid: 2499479
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>