[IMP] marketing_automation: replace create_text with add-label
before this commit: create_text was passed in node options and due to which it
was not parsed by translate.py, pass add-label as attribute on field so that
translate.py parse it and it is translated.

after this commit: create_text will be passed as field attribute instead of
node options.

task-1923433

closes odoo/enterprise#19418

Related: odoo/odoo#59713
Signed-off-by: Simon Genin (ges@odoo) <ges@odoo.com>