[FIX] {social}_push_notifications: allow internal users to access UTM campaigns
Before commit:
UTM campaigns are accessible for all internal users. But with the social
related moduels installed, few of the fields are added into `utm.campaign`
model and its views, which are accessible only to the social users/managers.

Because of that, whenever a user without basic social rights tries to access
the UTM campaings, AccessError is raised.

With this commit:
While adding fields to the campaign, we provide basic social group
("social.group_social_user") at field level  whenever necessary so it
doesn't raise AcessError.

For the views, we show the social related fields only to the users having
basic social rights. This way, we allow all the user to access campaign
views, and only show them the relevant data.

Task-2417993qq

X-original-commit: 20244e624215387300fd9778b12aff6ac21dc52c
Part-of: odoo/enterprise#21328