[FIX] project_enterprise: fix employee_id field does not exist in res.users
Before this commit, when we only install project_enterprise module, the
`employee_id` field is not defined in the `res.users` model. So, we have
a traceback since we check if the user has an employee_id to take his
resource calendar in `write` and `_calculate_planned_dates` methods in
`project.project` model.

This commit removes this check and gives the calendar of user if it is
set otherwise we take the default calendar set in the current company.

Related task: 2387984
Related PR: #16627

closes odoo/enterprise#19556

X-original-commit: 63b41300454656649d16ec8bb751cfbd1e781a26
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>