[FIX] account_bank_statement_import_camt: import stmt with different ccy
Suppose two different journals with the same bank account. The first one
is in USD, the second one in EUR. The import of a CAMT file with two
statements (one in USD, another in EUR) into one of the journals will
fail.

The user should be able to import the CAMT file once per journal

OPW-2558413

closes odoo/enterprise#19634

X-original-commit: 5f92b7c47c93d7798d60ca978cbb9c6b4787133c
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>