[IMP] documents, documents_hr*: bind document ownership to partner
Previously documents generated by hr apps (such as payslips) were
created with the employee's user as owner, but this does not allow for
databases which don't have an user per employee to store and share
documents for their employees only.

This commit adds the logic to bind the documents to the employee's
address and the user aswell.

Task ID: 2585039

closes odoo/enterprise#19452

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>