[IMP] industry_fsm_sale: add analytic account on task
Previously, all costs allocated to a task were associated, if set, to
the analytic account of their project.

This commit now uses the analytic account of the task and then
the one defined on the project (using the computed field which makes the
computation).

PR community: odoo/odoo#70527
PR enterprise: #19459

task-2458135 - part:3