[FIX] account_sepa: use bank account set on payment instead of partner
Before this commit, we were using the first bank account set on the partner associated
to the payment for which we are generating the XML file. We should use the bank account
set on the payment itself to avoid inaccuracies in case multiple bank accounts are set
on the partner.

closes odoo/enterprise#19567

X-original-commit: 3091b6d612d199e9e080edbf0abf7bce964196e5
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Alex Tuyls <alt-odoo@users.noreply.github.com>