[FIX] stock_barcode: tour fail often
In the `test_internal_picking_from_scratch_2` tour, at the end, it opens
a line form view and clicks on the product field. As this field is
readonly, it opens the product form view which isn't expected.
In some case, the test pass because it doesn't have the time to render
the product form view and so it's able to click on the save button and
then validate the picking.
But in case the form view is rendered before that, the test will fail.

Also, edits the comment as it didn't describe correctly what is done
and so, was more confusing than helping.

closes odoo/enterprise#19465

Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>