[FIX] web_gantt,project_enterprise: Allow users to groupby if no results
Problem: Traceback occurs when in projects - tasks - Gant View if you group by assigned to then project.
This only occurs if there is no tasks that are unassigned if 1 task is unassigned it works as it should.

Solution: Use an empty array by default instead of sending undefined

OPW-2587882

closes odoo/enterprise#21721

Forward-port-of: odoo/enterprise#19445
X-original-commit: eacdee067a4913e9d1c1b31066f0f2fba7e7b10d
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>