[IMP] project_enterprise: add connector home made component feature
This commit purpose is the introduction of new owl components allowing to represent tasks
dependencies in the web_gantt view.

The idea is to use an svg layer that will display the connectors and to allow the creation
of new ones through the use of dom selectors as being the allowed source and target for
the connectors.

task-2393690
closes PR: #18045