[IMP] documents_spreadsheet: add list view support
The user can now insert a list view in spreadsheet.

Two functions have been added:
- LIST.HEADER: Used to insert the columns name
- LIST: Used to insert the value. It takes an order as argument, which
represent the position of the records in the result of read_group, ordered
by the order defined on the list view when it was inserted.

These functions can be autofilled to re-build the list structure, or add
some records.

Like the pivot, it's possible to see the list attributes in a side panel,
re-insert a list and refresh the values.

This commit also modify the way the fetching of data is loaded. It's now
handle in a DataSource, which imply to update the pivot.

This commit also includes some updates of o-spreadsheet:
[IMP] style: support underline
[ADD] data_source
[REM] functions: remove async support
[IMP] menu_items: allow to define an icon class

Odoo-task-id 2516388

closes odoo/enterprise#20029

Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>