[FIX] l10n_in_reports: Remove duplicate GSTR-1 report lines
Current behavior before PR:
===========================
If child tax has multi-parents then the line in GSTR-1 is duplicated.

Desired behavior after PR is merged:
====================================
We simply remove duplicates by the union() on recordset so duplicates remove automatically

Additional info:
================
Before this commit odoo/odoo@7a7b5b5 when we search_read from orm then it gives uniquify list because there is one method _uniquify_list before that commit
and that method is removed so now we don't have uniquify list
so in this commit, I search first and uniquify using union and after read it so duplicate is removed automatically.

this problem is from SQL report but if we remove duplicate from SQL then it's performance killer so we do it from here.

opw-2577874

X-original-commit: 7ce3ccb70fe93a5f89065067fe0d3fe4a07be8ec