[FIX] project_timesheet_synchro: add session.js to manifest
Since 54e9527ec2fdf0467de08b372c7e891fedaa68ae, a new session module is
used to access the session info. The legacy session uses also this
module to access the session info. This means that if a module uses the
legacy session, the new session.js needs to be added to the manifest.

closes odoo/enterprise#19581

Related: odoo/odoo#73500
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>