[FIX] account_reports: reduce size of multi currencies revaluation
Even if reconciled, the vendor bill and the payment remain in the report
and the list becomes longer and longer as bills and payments are posted.
The report should remain an operational report and is not meant to be an
exhaustive list.
We only need the residual amounts once it is reconciled.

task-2584120

closes odoo/enterprise#20064

X-original-commit: 7b490cf25cdc96c830fc8a0962d51c61eaaf4088
Signed-off-by: William André (wan) <wan@odoo.com>