[IMP] sale_planning: add stat buttons on SO form view
With this commit, the hours_to_plan and the hours already planned are
displayed in the SO form view.

Those values are the sum of the hours to plan and hours planned on the
sols linked to the so.

PR community : odoo/odoo#75553
PR enterprise : #19647
PR upgrade : odoo/upgrade#2777

task-2559061 - part:4