[IMP] {,sale_}planning: improve copy previous week
With this commit, the employee calendar and planning are taken into
account to copy the previous week, so that the employee is available
when its slot is planned.

The algorithm is explained with further details in the code.

The copy previous week also take into account the remaining hours to
allocate for the slots depending on sol with plannable products. It
will copy slots until the remaining hours to schedule is 0.

PR community : odoo/odoo#75553
PR enterprise : #19647
PR upgrade : odoo/upgrade#2777

task-2559061 - part:8