[IMP] web_gantt,sale_planning: add planned hours progressbar
With this commit, a progressbar is displayed representing the number of
hours planned for an employee relatively to the number of hours the
employee is supposed to work, in the gantt view.
This progressbar is displayed on employee's, when there is a group by
employee.

PR community : odoo/odoo#75553
PR enterprise : #19647
PR upgrade : odoo/upgrade#2777

task-2559061 - part:9