[IMP] sale_planning: plan unscheduled slots from sol
With this commit, unscheduled slots linked to sol are assigned to the
best available employee following a priority.

This reuses the principle of the magnifying glass button on gantt view
to generate new slots (see previous commit for the same task), and try
to plan the unschedule slot, employee by employee, until one is found.

The priority list of employee to assign is meant to be easily overriden.

Tests are added to test the whole feature.

PR community : odoo/odoo#75553
PR enterprise : #19647
PR upgrade : odoo/upgrade#2777

task-2559061 - part:6