[FIX] account_reports: convert single mode into custom range
When consulting the Balance Sheet, if the user selects another date than
"Today", some lines (such as Current Year Earnings) are incorrect

To reproduce the error:
1. Create & Confirm 3 invoices:
    - INV01:
        - Date: today
        - Amount: $1000
    - INV02:
        - Date: today minus one month
        - Amount: $2000
    - INV03:
        - Date: today minus two months
        - Amount: $4000
2. Accounting > Reporting > US GAAP > Balance Sheet
3. Change the date filter: "End of Last Month"

Error: The amount of "Current Year Earnings" is $2000, it should be
$6000

When selecting "End of Last Month", a request is sent with a dates range
('Start of the month', 'End of the month') and the mode of the dates
filter is "single". The end date of the range will be the date used to
compute the values.

However, some lines are computed in "range" mode (this is the case for
Current Year Earnings). Therefore, it uses the dates range (the whole
month) to get the values. This explains why the amount $2000 is
displayed and why INV03 is ignored.

In such situation (computing a "range-mode" value in a "single-mode"
context), the start date should be defined with the beginning of the
fiscal year.

OPW-2554985

closes odoo/enterprise#19893

X-original-commit: a389761db6f831410f33a6741d840c4c903ac662
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>