[FIX] sale_amazon: avoid failing test on sale_amazon only build.
Since a2669c1 sale_amazon single module build was broken in test when
validating a picking

A first attempt to fix this was made in #19672. Unfortunattely this was
not enough since this error was also triggered in other tests and helped
to discover that this error wasn't leggit in some case: it should only
be triggered if sale_id is linked to an amazon_order_ref but the check
was triggered for all picking.

The first part of this commit checks that we have an amazon_picking
before checking if the module is installed.

The second part checks if we are in a test to avoid to trigger this
error on existing tests. This fix is crapy but easier to forwardport.

It looks like more of the logic could be put under the
`if amazon_pickings:` condition for performance reason but this will
target master.

closes odoo/enterprise#19757

X-original-commit: bd8685be8ffb002f0d6cca36f38b7d733f873ddb
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>