[IMP] mass_mailing_themes: adapt themes templates to new snippets
As many new snippets were added to `mass_mailing`, and many were
removed, this adapts the themes templates to these new snippets.

We also introduce the "VIP" and "Big News" themes to replace the "Resto"
and "Expo" themes. "Airmail" was removed.

The disabled templates ("resto", "expo" and "airmail") were not removed
in order to facilitate migration, but will be removed in the near
future.

task-2554899

Part-of: odoo/enterprise#21400
Co-authored-by: Elisabeth Dickinson <edi@odoo.com>
Co-authored-by: Antoine Guenet <age@odoo.com>