[IMP] web_enterprise,*: implement barcode scanning with Web APIs
In the process of replacing the native methods exposed in the mobile
apps, this commit implements the barcode scanning using the device's
camera through Web APIs.

The implementation is based on both `getUserMedia` to capture the video
stream from the device's camera (back if possible) and the barcode
detection using the Shape Detection API (specifically BarcodeDetector).

As this API is quite bleeding-edge and only supported in a subset of
browsers (cf. Chrome Android & Chrome Macos), a lightweight "polyfill"
(implementing only features used in our case) based on the Zxing for JS
library is used instead. This library is lazy-loaded only when the
polyfill is required to avoid requesting unnecessary assets.

Note: this feature requires a secureContext (aka. HTTPS) to be
available.

References:
- https://developer.mozilla.org/en-US/docs/Web/API/BarcodeDetector
- https://github.com/zxing-js/library
- https://developer.mozilla.org/en-US/docs/Web/API/MediaDevices/getUserMedia

Task ID: 2583914

Part-of: odoo/enterprise#20558
Co-authored-by: Pierre Paridans <app@odoo.com