[FIX] web_studio: fix incorrect offsets on high-dpi icons
This was caused by the fact that new icons were added to the image,
changing its dimensions. but the css fixing those dimensions to the
original ones was not adapted.

This commit fixes that by making the width "auto", allowing new icons to
be added without having to adapt the css.

closes odoo/enterprise#19873

X-original-commit: 5570459d56fa9ac200b8d04c1538069da6008abe
Signed-off-by: Simon Genin (ges@odoo) <ges@odoo.com>