[IMP] l10n_ch: improve financial reports
Improve the p&l and balance sheets by adding a control domain
and making sure all relevant accounts are found inside of it.

Also calculate the p&l results automatically in the balance sheet,
and add an action redirecting the user from the balance sheet to
the p&l on the p&l line.

Task id #2585627

closes odoo/enterprise#19830

Signed-off-by: Florian Gilbert <flg@odoo.com>