[IMP] helpdesk: improve the team team kanban card
Purpose of the commit is to improve the team kanban card
layout on dashboard so user can easiy access the tickets.

So in this commit added link in the top right corner for:
Avg Open Hours, Tickets Closed, Success Rate, Customer Satisfaction
and added the button at the bottom for: Open, Unassigned, Urgent, Failed
Also added the butger menu in kanban card.

task-2531483

closes odoo/enterprise#20075

Related: odoo/upgrade#2718
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>