[FIX] timesheet_grid: keep view order when modifying with studio
The Timesheets > To Validate menus are created with a server action
instead of directly being `act_window` but they copy most of their
parameters from an existing view.
When modifying the default view using studio, the change does not apply
to the To Validate views, this commit aims to fix that issue.
This is done by sorting the views according to the action's `view_mode`
field.

Closes odoo/enterprise#19886

Task ID: 2612215

closes odoo/enterprise#21853

X-original-commit: b25a6e10d586c650329200fea3aa7f95b92553a8
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>