[FIX] helpdesk: fix list view not being present in ratings
When clicking on rating in the helpdesk dashboard, the list view would
not be available until the user reloads the page. This is because the js
client expects `list` instead of `tree` in some cases and the action was
handled differently than on simple buttons.
Now `tree` will be replaced by `list` in the list of views for the
action.

Also fixes the view showing ratings unrelated to helpdesk after
refreshing.

Closes odoo/enterprise#19890

Task ID: 2612307

closes odoo/enterprise#20320

X-original-commit: 954243bf0e1e1763eb4cd7ad713567dd1c4d7d63
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>