[FIX] sale_subscription: fetch cust ref on invoice generation
How to reproduce the problem:
- Install the sale_subscription module
- Create a SO with a subscription product
(e.g. 'Office Cleaning Subscription (Monthly)')
- In the 'Other info' tab of the SO,
write something in the Customer Reference field
- Save & Confirm the SO. Click on the Subscription button.
- (Activate the debug mode) Click on Generate Invoice.
- The Payment Reference of the created invoie is empty.
-> It should correspond to the Customer Reference of the SO

Cause of the problem : the value was simply not fetched

Solution : the reference is fetched from the sale order that generated
the Subscription. If multiple Sale Orders are linked to
the same subscription, the references are joined into one.

opw-2525643

closes odoo/enterprise#20290

X-original-commit: 1423c4b068251724b95a5e548b5584344e623329
Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>
Signed-off-by: prro-odoo <proose@users.noreply.github.com>