[FIX]: mail_enterprise: The chatter does not take all the space for the big screen
The issue appear in the invoices view when the record has at least an
attachment and `account_invoice_extract` is not installed. If the module
is installed, an xpath replaces the part displaying the preview of
attachments and modifies the domain of the modifier.
The classname `o-aside` are not applied because the attachment preview
exists. The attachment preview is created because there are attachments
but is not inserted because the container is invisible (applied domain).

The fix consists in checking if the preview attachment container is
visible to create the preview.

In community:
```xml
<div class="o_attachment_preview"
    attrs="{'invisible': ['|',
        ('move_type', 'not in', ('out_invoice', 'out_refund', 'in_invoice', 'in_refund')),
        ('state', '!=', 'draft')]}" />
```
In enterprise:
```xml
<xpath expr="//div[hasclass('o_attachment_preview')]" position="replace">
   <div class="o_attachment_preview" options="{'preview_priority_type': 'pdf'}"/>
</xpath>
```

opw-2517726
opw-2547527

closes odoo/enterprise#19946

X-original-commit: 3a0506ccd49bb5db92535b20436c8c3055e1ff07
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>