[FIX] hr_appraisal_survey: Appraisal multi-company issue
To replicate an issue:

Install website. Have multiple companies. Set company A for the website.
Check the company B from the multi-company widget.
Now create appraisal for employee (that belongs to B company). Ask Feedback
for the employee. Click on the 'Feedback Survey' and we get
- Appraisal multi-company - restriction.

After this fix, we won't get the same issue

task - 2528209

closes odoo/enterprise#20408

X-original-commit: 57057b9df9c09c0d2d0b10e3a191a58893f1caf1
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>