[FIX] l10n_mx_edi: Don't request a signature for cancelled invoices
If you cancel a Mexican invoice and then attempt to print it, a red
message will show on the invoice saying a signature is required. This
doesn't make sense since the invoice was cancelled. This fix doesn't
display that message if the invoice is in the cancelled state.

In V13 and earlier the logic was different: the message would display
if the folio fiscal was expected but not filled in. In the case of
cancelled invoices in V13, the folio fiscal would remain filled in, so
the message didn't trigger. In V14 the folio fiscal is cleared in that
case, and the trigger logic was rewritten.

opw-2582779

closes odoo/enterprise#19998

X-original-commit: f03dac515596fe3a1acec3e958cc0a1893283897
Signed-off-by: Josse Colpaert <jco@openerp.com>