[IMP] account_reports: process folded lines in batch in render function
Currently each call to account_reports.render() loop through
each js_account_report_foldable lines and calls this.fold on it.
In turn, this.fold searches this.$el to retrieve the 'tr' objects
to hide in the report. This means that currently the rendering
process is more or less in O(n²) with n the number of lines in the report.

To improve that, call this.$el.getElementsByTagName('tr') only once
and then apply the 'fold' modifications to the correct lines.
This greatly reduces the time taken by the browser to render
a given account_reports when the number of lines gets bigger (>2000).

closes odoo/enterprise#22843

X-original-commit: 420d64fc76ad54da031875bf6514a787218197b6
Signed-off-by: William André (wan) <wan@odoo.com>