[FIX] l10n_ar_reports: document type wrong codes
When we are computing the invoices to export the vat book, we need to
check if the document type is a "liquidación". At the moment of that
check we were comparing with the wrong document types.

Another related fix is that if the doc type is "liquidación" we need to
report the vat, and if not we just report a string of eleven 0's.

There is no way to replicate the bug in Odoo since the purchase vat book
is exported ok, but when you go to upload in Afip, you get an error.

closes odoo/enterprise#21146

X-original-commit: 9f02c3bdc6ff7e66ab5fddecafe1f9fe14b21e5e
Signed-off-by: Josse Colpaert <jco@openerp.com>