[IMP] sign: Change the requester of the sign request
In the signing page of a sign.request, show the sign.request's sender(creator)'s info on the top as the requester

before this commit:
If I have a template whose responsible is user X
and I found that template usefull. I'll reuse it to create a new sign request.
On the new sign request, the requestor (display top left) is still the responsible of the template, user X.
But for the end user, if he have any question, he need to ask the people that contacted him, that requested the signature.
The reference user should be the one that create the signe request.

closes odoo/enterprise#19966

Taskid: 2614061
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>