[IMP] social: simplify the "description" field on social.media
The description field on a social media currently stores HTML content.
Which feels completely unnecessary since these records are only meant to be
created as data and the content of this field is very simple (one regular
sentence).

This commit simplifies the description by storing regular text instead.

(This also allows to solve an issue of rendering introduced by
9f4b0a05149e1e0d1b11513c75801736779eceae ).

We also took this opportunity to rename it to "media_description" to ease
greping.

Task-2613240
UPG PR odoo/upgrade#2695

closes odoo/enterprise#19985

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>