[FIX] helpdesk_sale_timesheet: fix traceback when mass editing
This commit fixes a traceback when mass editing sale_line_id in
helpdesk.ticket's tree view. It is caused by the need of
`commercial_partner_id` in the domain of `sale_line_id` not being
present in the view. We will now not support the mass editing of
`sale_line_id` instead of leaving it editable since it does not make a
lot of sense to mass edit a sale order line anyway.

Closes odoo/enterprise#19987

Task ID: 1601707

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>