[FIX] industry_fsm_sale_report: fix multi company issue
Fixes a multi company issue with the `worksheet_template_id` field on
`product.template`, the field depends on a company_dependent field and
it would not retrigger the compute upon changing the current active
company. The field is now a company_dependent field aswell and is
recomputed when writing/creating records.

Closes: odoo/enterprise#20007

Task ID: 2613947

Related: odoo/upgrade#2702
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>