[FIX] social: fix social form views "Engagement" stat-buttons
Since the latest OWL refactoring, stat-buttons "button" DOM element in form
views cannot be defined without an associated action (it will crash on click).

In the social module, we have 2 engagement fields for statistics, on
social.posts and on social.campaigns.
The engagement statistics are displayed using the style of a stat-button in
those models' form views, but have no associated action.

To fix the issue, we now mark the buttons as "disabled" so the click action
will not trigger anything.

Task-2618108

closes odoo/enterprise#20049

Related: odoo/odoo#75158
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>