[FIX] mrp_workorder: select product in manufacturing quality point
When selecting a manufacturing picking_type, it was impossible to select
a product unless a workorder operation was selected. It also removed all
products selected when doing so.
Allow products to be selected again when no workorder operation is
selected, while still restricting it when a workorder is set.

Task-2524922

Part-of: odoo/enterprise#20250