[FIX] timesheet_grid: fix the usability improvement for timesheet settings
Purpose of this commit is to display the value of
reminder_user_delay and reminder_manager_delay fields
center in the setting of the timesheet app.

So, In this commit add a text-center class for that fields
in XML of the setting in the timesheet app.

closes odoo/enterprise#20135

Taskid: 2616165
Related: odoo/odoo#74849
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>