[FIX] social_{crm, sale}: allow internal users to access UTM campaign
Before this commit users were not able to access social.post and
utm.campaign due to the following compute fields
1) leads_opportunities_count
2) sale_quotation_count
3) sale_invoiced_amount
As we are computing them without checking whether user has its rights or not.

So, in this commit we checks whether user has 'sales_team.group_sale_salesman'
group or not. If user does not have this group than the values will not be
computed.

Task-2417993

closes odoo/enterprise#21328

X-original-commit: cdd72fb2c20f7b08365a7ea134a144f213de98ae
Related: odoo/odoo#77558
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>