[REM] mail_github: remove the module
The mail_github module does not seem to be particularly used by our customers.
It is not installed on our prod either even though we actively use github. It is
difficult to find it because it cannot be installed from the settings. At last,
it caters a very specific segment of customers. All in all, it does not seem
worth it to continue maintaining this module.

task-2369600

closes odoo/enterprise#20147

Related: odoo/upgrade#2724
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>