[REF] sign: refactor update_from_pdfviewer
[1] separate duplicate
decouple the logic "duplicate" and "update_from_pdfviewer"

before this commit:
update_from_pdfviewer is used for duplicating templates
A more intuitive and general approach is
1. decouple the log "duplicate" and "update_from_pdfviewer"
2. override the `copy` function of the sign.template

[2] no recreate
new sign items of a template will only be created once in database

before this commit: when you open a sign template, and add some new sign items in to the template.
Every time the (auto) 'save' is triggered, these new added sign items will be removed in the database
and recreated again with a different id. This approach is strange and should be avoided.

[3] better stability
use copy() to improve the stability.

Before this commit:
update_from_pdfviewer is quite special, because it will write to all sign items no matter if their data are really changed or not.
So, if a user updates the sign template twice within in a short period(e.g. change the location of a sign item twice).
Or `saveTemplate` function is called twice within one user action.
Because of some network issues, the two requests of these two RPC come to the server almost at the same time.
The two RPCs may want to write to the same record with the same value at the same time.

As a result,
In database, there will be an error `not serialize access due to concurrent update`
In server, our strategy for serialization failure is `retry`
Since the input of the update_from_pdfviewer has been changed by itself and these data cannot be reverted, an error will be raised and cause a traceback for user.

[4] update sign items

After this commit:
* update sign items when changed, so the number of SQL queries is reduced
* The transaction_id is used to prevent recreating/mis-deleting sign items when the network is not stable
    Note: transaction_id == 0 means:
    1. the sign.item is not created by pdfviewer (maybe be created by other python code) or
    2. The transaction for `create` has been confirmed by the pdf_viewer (it is updated after created)
* Easier to identify sign items in js
    We can always use 'itemId' to identify every sign item
    1. sign items existing in the database:
        positive itemId
    2. sign items new created (haven't been uploaded to server):
        negative random itemId. This itemId is also used as the transaction_id while uploading

Before this commit:
All sign items are updated when the customer updates the template

[5] style

remove model.api
create new sign items in batch

Before this commit:
the update_from_pdfviewer uses template_id as a parameter,
but there is no reason for one template to update_from_pdfviewer for another template
So it is inconvenient to reuse this method from backend code
new sign items are created in a for loop.

Taskid: 2659342
Part-of: odoo/enterprise#22813