[IMP] hr_payroll: improve work entry type form view
currently, the unpaid structures type input field doesn't
take the full width. it happens because the div class is added
by the xpath in the unpaid group.

so this commit fixes the issue by removing the unnecessary xpath
so that the unpaid structures type input field takes a full width

TaskID-2611955

closes odoo/enterprise#20154

Signed-off-by: Kevin Baptiste <kba@odoo.com>