[IMP]l10n_dk_reports: prev. year earnings calc.
The aim of this commit is to allow Odoo to calculate the previous year
unallocated earnings in order to get a correct report.

Before this commit, the result of the previous year wasn't taken into
account which resulted in an unbalanced report if the previous year
earnings weren't allocated. (Nulkontrol != 0)

Now, the report calculates the current year earnings and the
previous years earnings.

Also, previous to this commit, the passiv was displayed with negative
numbers. It is now displayed with positive numbers.

closes odoo/enterprise#21564

X-original-commit: 9b955ad0a272359083cec09bbe2c8ac664285d8c
Signed-off-by: Josse Colpaert <jco@openerp.com>
Signed-off-by: Brice Bartoletti <Levizar@users.noreply.github.com>