[IMP] l10n_fr_reports: Round lines of tax report and closing entries to the nearest euro
French tax reports require that each line is rounded to the nearest euro.
This requires that a line is added to the closing entries to account for the
differences from rounding.

In order to implement the rounding inside l10n_fr_reports module, the
behaviour of the function _compute_vat_closing_entry must be adapted.
To avoid completely redefining it inside of l10n_fr_reports, a function
is added called _postprocess_vat_closing_entry_results, which is called
from within _compute_vat_closing_entry.

The _postprocess_vat_closing_entry_results function must take as its
input the results of the vat closing entry query defined inside of
_compute_vat_closing_entry. These results contain the information used
to generate the lines of the closing entry, so adapting them allows us
to apply the rounding. It taks as its arguments the company, the options
and the results, and returns the results with the rounding difference
line applied.

The fields:
- l10n_fr_rounding_difference_profit_account_id -
- l10n_fr_rounding_difference_loss_account_id   -
are added to the res.config.settings model (and the res config settings
view), to allow the user to configure the accounts which are utilised
for loss and profit accumulated as a result of rounding the entries.

A function called:
_l10n_fr_get_lines_to_round_closing
Is added, which returns two lists (one for sales taxes, one for purchase
taxes) of the xml_ids of the account.tax.report.line that are to be used
in the calculation of the rounding difference.

Tests are included to test the behaviour of different rounding cases

task-2610752

Part-of: odoo/enterprise#20245