[IMP] planning: set working days count as float
`working_days_count` should be evaluated as a Float instead of an
Integer, otherwise it could lead to incorrect result: suppose a user
adds two slots on the same day and for the same employee. For now,
`working_days_count` is an integer and its value is rounded up, so this
field value for each slot will be 1 and therefore, even if the employee
worked one day, the sum of `working_days_count` will be 2.

From now on, `working_days_count` are stored as not-rounded floats (to
remain as accurate as possible in the calculations). Note that, on
client side, the displaying of this field will be automatically rounded:
https://github.com/odoo/odoo/blob/e89b626f5411b7836030c6daa940b41f65747843/addons/web/static/src/legacy/js/fields/field_utils.js#L175-L174

Linked to OPW-2496638

closes odoo/enterprise#20441

Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>