[IMP] helpdesk: Improve team visibility mechanism
The purpose:
The way portal and internal users are granted the reading access to the records,
differs in the helpdesk from the other modules, which is a bit inconsistent from the users point of view.

In this task, the above mentioned behavior is aligned to how it's done in the module - project.

task-2590082

closes odoo/enterprise#20190

Related: odoo/upgrade#2758
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>