[FIX] helpdesk: fix the traceback on pivot view of ticket report
Currently when goto the ticket analysis reporting >> pivot view and
add the measure `close hours` gives traceback.
traceback is due to the use of timesheet_uom widget which is defined
in the hr_timesheet and when only helpdesk installed at that time it
will generate an error.

So in this commit use the float_time widget instead of the timesheet_uom.

Second, when we create a ticket, and on the form view you can see the
stat button showing ticket -1 if not ticket found that linked to same customer

so to fix the -1 count if any tickets found for the same customer then do the -1
else count as 0

closes odoo/enterprise#20174

Taskid: 2619425
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>