[FIX] website_helpdesk_forum: fix the html tag display in forum post
Currently, go to helpdesk > configuration > helpdesk team > enable the
'help center' feature.
create a ticket > add a description > share on the forum
the html tags are display in forum post because previously the description
field was text while now it's converted to the html.

so while creating the forum post from the ticket do not convert the html
description field data into html again using the plaintext2html.
directly pass the html as it is as description field is now html field.

closes odoo/enterprise#20340

Taskid: 2602591
X-original-commit: a2b4958e5bc6dff2d9ee65ae6f4f2c0c7a19fcb3
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Mahendra Barad <mba-odoo@users.noreply.github.com>